Remove old timeline
