+++ src/timeline/TimelineViewManager.h
@@ -1,69 +1,80 @@
-/*
- * nheko Copyright (C) 2017  Konstantinos Sideris <siderisk@auth.gr>
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
 #pragma once
 
+#include <QQuickView>
+#include <QQuickWidget>
 #include <QSharedPointer>
-#include <QStackedWidget>
+#include <QWidget>
 
-#include <mtx.hpp>
+#include <mtx/responses.hpp>
 
+#include "Cache.h"
+#include "Logging.h"
+#include "TimelineModel.h"
 #include "Utils.h"
 
-class QFile;
+// temporary for stubs
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wunused-parameter"
 
-class RoomInfoListItem;
-class TimelineView;
-struct DescInfo;
-struct SavedMessages;
+class MxcImageProvider;
+class ColorImageProvider;
 
-class TimelineViewManager : public QStackedWidget
+class TimelineViewManager : public QObject
 {
         Q_OBJECT
 
+        Q_PROPERTY(
+          TimelineModel *timeline MEMBER timeline_ READ activeTimeline NOTIFY activeTimelineChanged)
+
 public:
-        TimelineViewManager(QWidget *parent);
-
-        // Initialize with timeline events.
-        void initialize(const mtx::responses::Rooms &rooms);
-        // Empty initialization.
-        void initialize(const std::vector<std::string> &rooms);
-
-        void addRoom(const mtx::responses::JoinedRoom &room, const QString &room_id);
-        void addRoom(const QString &room_id);
+        TimelineViewManager(QWidget *parent = 0);
+        QWidget *getWidget() const { return container; }
 
         void sync(const mtx::responses::Rooms &rooms);
-        void clearAll() { views_.clear(); }
+        void addRoom(const QString &room_id);
 
-        // Check if all the timelines have been loaded.
-        bool hasLoaded() const;
+        void clearAll() { models.clear(); }
 
-        static QString chooseRandomColor();
+        Q_INVOKABLE TimelineModel *activeTimeline() const { return timeline_; }
+        void openImageOverlay(QString mxcUrl,
+                              QString originalFilename,
+                              QString mimeType,
+                              qml_mtx_events::EventType eventType) const;
+        void saveMedia(QString mxcUrl,
+                       QString originalFilename,
+                       QString mimeType,
+                       qml_mtx_events::EventType eventType) const;
+        Q_INVOKABLE void cacheMedia(QString mxcUrl, QString mimeType);
+        // Qml can only pass enum as int
+        Q_INVOKABLE void openImageOverlay(QString mxcUrl,
+                                          QString originalFilename,
+                                          QString mimeType,
+                                          int eventType) const
+        {
+                openImageOverlay(
+                  mxcUrl, originalFilename, mimeType, (qml_mtx_events::EventType)eventType);
+        }
+        Q_INVOKABLE void saveMedia(QString mxcUrl,
+                                   QString originalFilename,
+                                   QString mimeType,
+                                   int eventType) const
+        {
+                saveMedia(mxcUrl, originalFilename, mimeType, (qml_mtx_events::EventType)eventType);
+        }
 
 signals:
         void clearRoomMessageCount(QString roomid);
-        void updateRoomsLastMessage(const QString &user, const DescInfo &info);
+        void updateRoomsLastMessage(QString roomid, const DescInfo &info);
+        void activeTimelineChanged(TimelineModel *timeline);
+        void mediaCached(QString mxcUrl, QString cacheUrl);
 
 public slots:
         void updateReadReceipts(const QString &room_id, const std::vector<QString> &event_ids);
-        void removeTimelineEvent(const QString &room_id, const QString &event_id);
         void initWithMessages(const std::map<QString, mtx::responses::Timeline> &msgs);
 
         void setHistoryView(const QString &room_id);
+        void updateColorPalette();
+
         void queueTextMessage(const QString &msg);
         void queueReplyMessage(const QString &reply, const RelatedInfo &related);
         void queueEmoteMessage(const QString &msg);
@@ -90,9 +101,17 @@ public slots:
                                uint64_t dsize);
 
 private:
-        //! Check if the given room id is managed by a TimelineView.
-        bool timelineViewExists(const QString &id) { return views_.find(id) != views_.end(); }
+#ifdef USE_QUICK_VIEW
+        QQuickView *view;
+#else
+        QQuickWidget *view;
+#endif
+        QWidget *container;
+        TimelineModel *timeline_ = nullptr;
+        MxcImageProvider *imgProvider;
+        ColorImageProvider *colorImgProvider;
 
-        QString active_room_;
-        std::map<QString, QSharedPointer<TimelineView>> views_;
+        QHash<QString, QSharedPointer<TimelineModel>> models;
 };
+
+#pragma GCC diagnostic pop
